Move enums to separate file
